Server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
yfinance
scikit-learn
python-dotenv
